+++ public/export_selected_bundle.php
@@ -16,7 +16,6 @@ if ($selected_export_cli_job) {
 error_reporting(E_ALL & ~E_DEPRECATED);
 ini_set('display_errors', '0');
 
-ini_set('memory_limit', '512M');
 set_time_limit(600);
 ignore_user_abort(true);